Update .NET SDK
Update to the latest .NET 6 RC1 SDK.
Fix benchmarks.
