+++ tests/API.Benchmarks/API.Benchmarks.csproj
@@ -1,6 +1,8 @@
-<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
     <Description>Benchmarks for the API.</Description>
+    <!-- HACK Workaround for https://github.com/dotnet/sdk/issues/20444 -->
+    <LangVersion Condition="'$(BuildingInsideVisualStudio)' == 'true'">preview</LangVersion>
     <NoWarn>$(NoWarn);CA2007;CA2234;SA1600</NoWarn>
     <OpenApiGenerateDocuments>false</OpenApiGenerateDocuments>
     <OutputType>Exe</OutputType>
